backend/requirements.txt
Flask
Flask-Cors
qiskit
qiskit-aer
google-generativeai
python-dotenv
